Rename  ccloudResourceLoaderUtils -> udfSystemCatalogQuery.
